More small cleanup
isort imports of changed files
move sanity_check to builtin.system_init
mark some no_doc modules
