+++ mathics/builtin/__init__.py
@@ -5,33 +5,25 @@ Mathics Built-in Functions and Variables.
 Mathics has over a thousand Built-in Functions and variables, all of which are defined here.
 """
 
-from typing import Optional
 import glob
 import importlib
 import inspect
+import os.path as osp
 import pkgutil
 import re
-import os.path as osp
+from typing import List, Optional
 
 from mathics.builtin.base import (
     Builtin,
-    SympyObject,
     Operator,
     PatternObject,
+    SympyObject,
+    mathics_to_python,
 )
-
+from mathics.builtin.system_init import sanity_check
 from mathics.core.pattern import pattern_objects
-
 from mathics.settings import ENABLE_FILES_MODULE
 from mathics.version import __version__  # noqa used in loading to check consistency.
-
-from typing import List
-
-# Set this to True to print all the builtins that do not have
-# a summary_text. In the future, we can set this to True
-# and raise an error if a new builtin is added without
-# this property or if it does not fulfill some other conditions.
-RUN_SANITY_TEST = False
 
 # Get a list of files in this directory. We'll exclude from the start
 # files with leading characters we don't want like __init__ with its leading underscore.
@@ -75,9 +67,9 @@ def contribute(definitions):
         if name != "System`MakeBoxes":
             item.contribute(definitions)
 
+    from mathics.core.definitions import Definition
     from mathics.core.expression import ensure_context
     from mathics.core.parser import all_operator_names
-    from mathics.core.definitions import Definition
 
     # All builtins are loaded. Create dummy builtin definitions for
     # any remaining operators that don't have them. This allows
@@ -161,21 +153,6 @@ def name_is_builtin_symbol(module, name: str) -> Optional[type]:
     return module_object
 
 
-def sanity_check(cls, module):
-    if not RUN_SANITY_TEST:
-        return True
-
-    if not hasattr(cls, "summary_text"):
-        print(
-            "In ",
-            module.__name__,
-            cls.__name__,
-            " does not have a summary_text.",
-        )
-        return False
-    return True
-
-
 # FIXME: redo using importlib since that is probably less fragile.
 exclude_files = set(("codetables", "base"))
 module_names = [
@@ -249,7 +226,6 @@ for module in modules:
                 builtins_by_module[module.__name__].append(instance)
 
 mathics_to_sympy = {}  # here we have: name -> sympy object
-mathics_to_python = {}  # here we have: name -> string
 sympy_to_mathics = {}
 
 builtins_precedence = {}
